requirement.txt
torch
pandas
numpy
glob2
pillow
boto3